backend/src/services/ai-services/requirements.txt
tensorflow-cpu==2.11.0